archives/person_module_scraper/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
black==21.5b2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.6
Django==3.1.7
djangorestframework==3.12.2
fake-useragent==0.1.11
httmock==1.4.0
http-request-randomizer==1.3.2
idna==2.10
loguru==0.5.3
mypy-extensions==0.4.3
numpy==1.20.1
pandas==1.2.3
password-validator==1.0
pathspec==0.8.1
Pillow==8.1.2
piplapis-python==5.2.2
psutil==5.8.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
pymongo==3.11.3
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.1
python-environ==0.4.54
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
validate-email==1.3
win32-setctime==1.0.3
